Move projects service to projects-list component
